Extract TestSceneFocusedOverlayContainer
